still trying to figure out why it's acting weird
I would assume that by making a new image based of the last edit would return it to the original position and that by inputing a number it will shift it to that position, but instead it's shifting it by that number instead.  Wait a sec, that gives my an idea.