Holy flip! make3D() is working!
I figured out what was wrong.  Originally, temp was making a copy of activeImage (which is the image that is currently displayed).  The thing is, when you call setCurrentImage, it will set activeImage to be what ever the parameter is, in this case the already altered image.  Before, it was essentially shifting an image that was already shifted instead of shifting what was there before.  My logic was right originally, it's just that the image wasn't correct.