Started rewrite for EditStack
I realized just how hard it is to complete change what your original line of thinking was.
Instead or adding an image to the stack, it will need to now add a profile to the stack, which will be hard to do without crossing over the model with the view.  That just gave me an idea, what if instead I make a method for the controller that makes a new profile!